Update Game and GameDetail models to use reusable formatting functions for common data
